+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Categories list page
 - Categories details page
 - Set category color
+- Add priority switch in habit, for more prominent display
 
 ## [0.8.285] - 2023-03-19
 ### Added: